requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
sqlalchemy>=2.0
sentry-sdk[fastapi]>=2.0
httpx>=0.27
python-dotenv>=1.0
../taskpilot-shared
